feat(forms): conditional rendering and UI enhancements
i) In the allocate page, create conditional rendering of the form input depending on the transfertype. This ensures that only relevant fields are displayed based on the selected transfertype.
ii) In the Results page, added organisation and final buyer name for better user awareness. This provides more context to the user regarding the results displayed.
iii) In the transactions page, removed txn id and added final buyer, txn no, and txn data columns in the table and PDF. This improves the clarity and relevance of the transaction information presented.